First upload of data processing script
Uses FFT to approximate the original signal based on either (1) a specified percentage of the most significant frequencies or (2) a specified number of the most important frequencies. Needs the original acceleration data (.mat file) to be in the same directory. Saves the cleaned data to a .mat file in the working directory.

Order of saved data:
- Linear x
- Linear y
- Linear z
- Angular x
- Angular y
- Angular z