Flight Software Offshoot for Attitude Control Test
This is the code used for testing on 3/29.
The main purpose of this code is to trick the flight computer into "terminalGuidance" (machineState = TERMINALGUIDANCE). This is also coupled with a command to fire the descent motor, and the resetting of values that must start at some 0. The code in the main loop all runs every loop, but within the switch statement (state machine) only terminalGuidance should be running at the time of the test. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!Do not upload any code to the flight computer until monday when the flight computer has been taken out of the vehicle and the test data has been offloaded!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! I also would be careful about changing any code that deals with flash memory, which should not be needed in the first place.